lab21/requirements.txt
pip
fastapi
uvicorn
Pillow
Jinja2
gdown
python-multipart
tensorflow<2.11
